LayoutTests/fast/forms/input-number-commit-valid-only-expected.txt

Type=number field should not accept invalid numbers though a user can type such strings

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "512"
PASS input.value is "512"
PASS input.value is ""
PASS successfullyParsed is true

TEST COMPLETE
